Move runAcceptTerminal to run_common.go
